Move EventsAndDenseTriggers lib to ParallelAlgorithms
