+++ src/ControlSystem/Trigger.hpp
@@ -12,11 +12,11 @@
 #include "ControlSystem/CombinedName.hpp"
 #include "ControlSystem/FutureMeasurements.hpp"
 #include "ControlSystem/Metafunctions.hpp"
-#include "Evolution/EventsAndDenseTriggers/DenseTrigger.hpp"
 #include "IO/Logging/Verbosity.hpp"
 #include "Parallel/ArrayComponentId.hpp"
 #include "Parallel/Callback.hpp"
 #include "Parallel/GlobalCache.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/DenseTrigger.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/GetOutput.hpp"
 #include "Utilities/Gsl.hpp"
