+++ src/Evolution/CMakeLists.txt
@@ -61,7 +61,6 @@ add_subdirectory(Conservative)
 add_subdirectory(Deadlock)
 add_subdirectory(DgSubcell)
 add_subdirectory(DiscontinuousGalerkin)
-add_subdirectory(EventsAndDenseTriggers)
 add_subdirectory(Executables)
 add_subdirectory(Imex)
 add_subdirectory(Initialization)
