+++ src/ParallelAlgorithms/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(Actions)
 add_subdirectory(ApparentHorizonFinder)
 add_subdirectory(Amr)
 add_subdirectory(Events)
+add_subdirectory(EventsAndDenseTriggers)
 add_subdirectory(EventsAndTriggers)
 add_subdirectory(Initialization)
 add_subdirectory(Interpolation)
