+++ src/ParallelAlgorithms/EventsAndDenseTriggers/DenseTrigger.hpp
