+++ src/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/CMakeLists.txt
