+++ src/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/Times.hpp
@@ -7,10 +7,10 @@
 #include <optional>
 #include <pup.h>
 
-#include "Evolution/EventsAndDenseTriggers/DenseTrigger.hpp"
 #include "Options/Options.hpp"
 #include "Options/ParseOptions.hpp"
 #include "Options/String.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/DenseTrigger.hpp"
 #include "Time/TimeSequence.hpp"
 #include "Utilities/Serialization/CharmPupable.hpp"
 #include "Utilities/TMPL.hpp"
