+++ src/ParallelAlgorithms/EventsAndDenseTriggers/EventsAndDenseTriggers.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/EventsAndDenseTriggers/EventsAndDenseTriggers.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/EventsAndDenseTriggers.hpp"
 
 #include <algorithm>
 #include <pup.h>
