+++ src/ParallelAlgorithms/EventsAndDenseTriggers/Tags.hpp
@@ -4,8 +4,8 @@
 #pragma once
 
 #include "DataStructures/DataBox/Tag.hpp"
-#include "Evolution/EventsAndDenseTriggers/EventsAndDenseTriggers.hpp"
 #include "Options/Options.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/EventsAndDenseTriggers.hpp"
 #include "Utilities/Serialization/Serialize.hpp"
 #include "Utilities/TMPL.hpp"
