+++ tests/Unit/Evolution/CMakeLists.txt
@@ -6,7 +6,6 @@ add_subdirectory(Ader)
 add_subdirectory(BoundaryConditions)
 add_subdirectory(DgSubcell)
 add_subdirectory(DiscontinuousGalerkin)
-add_subdirectory(EventsAndDenseTriggers)
 add_subdirectory(Imex)
 add_subdirectory(Initialization)
 add_subdirectory(Particles)
