+++ tests/Unit/Helpers/Evolution/CMakeLists.txt
@@ -1,7 +1,6 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-add_subdirectory(EventsAndDenseTriggers)
 add_subdirectory(DgSubcell)
 add_subdirectory(DiscontinuousGalerkin)
 add_subdirectory(Imex)
