+++ tests/Unit/Helpers/ParallelAlgorithms/CMakeLists.txt
@@ -2,6 +2,7 @@
 # See LICENSE.txt for details.
 
 add_subdirectory(Events)
+add_subdirectory(EventsAndDenseTriggers)
 add_subdirectory(Interpolation)
 add_subdirectory(LinearSolver)
 add_subdirectory(NonlinearSolver)
