+++ tests/Unit/Helpers/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/TestTrigger.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Helpers/Evolution/EventsAndDenseTriggers/DenseTriggers/TestTrigger.hpp"
+#include "Helpers/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/TestTrigger.hpp"
 
 namespace TestHelpers::DenseTriggers {
 PUP::able::PUP_ID TestTrigger::my_PUP_ID = 0;  // NOLINT
