+++ tests/Unit/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/CMakeLists.txt
