+++ tests/Unit/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/Test_Filter.cpp
@@ -9,16 +9,16 @@
 #include <string>
 
 #include "DataStructures/DataBox/DataBox.hpp"
-#include "Evolution/EventsAndDenseTriggers/DenseTrigger.hpp"
-#include "Evolution/EventsAndDenseTriggers/DenseTriggers/Filter.hpp"
-#include "Evolution/EventsAndDenseTriggers/DenseTriggers/Times.hpp"
 #include "Framework/TestCreation.hpp"
 #include "Framework/TestHelpers.hpp"
-#include "Helpers/Evolution/EventsAndDenseTriggers/DenseTriggers/TestTrigger.hpp"
+#include "Helpers/ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/TestTrigger.hpp"
 #include "Options/Protocols/FactoryCreation.hpp"
 #include "Options/String.hpp"
 #include "Parallel/GlobalCache.hpp"
 #include "Parallel/Tags/Metavariables.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/DenseTrigger.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/Filter.hpp"
+#include "ParallelAlgorithms/EventsAndDenseTriggers/DenseTriggers/Times.hpp"
 #include "ParallelAlgorithms/EventsAndTriggers/Trigger.hpp"
 #include "Time/Slab.hpp"
 #include "Time/Tags/Time.hpp"
